Migrate to aws-sdk-go-v2
I was having a lot of issues getting this to work correctly with SSO and
migrating to v2 of the SDK fixed the issue. I also ended up writing a few tests
around the pricing parsing since I ran into issues trying to get that to work
correctly with v2.